backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.5
sendgrid==6.10.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
PyMuPDF==1.23.8
python-docx==1.1.0
langchain==0.0.340
langchain-openai==0.0.2
httpx==0.25.2
